Deprecate old IC transform (#62)
* deprecate old ic transform

* new_ic -> ic
