+++ bgtorch/nn/flow/crd_transform/__init__.py
@@ -1,5 +1,2 @@
 from .pca import *
-from .new_ic import *
-
-# To be removed:
-from .ic import MixedCoordinateTransform
+from .ic import *
